--- pages/mes-jeunes/milo.tsx
@@ -1,16 +1,0 @@
-import { withTransaction } from '@elastic/apm-rum-react'
-
-function Milo() {
-  return null
-}
-
-export const getServerSideProps = () => {
-  return {
-    redirect: {
-      destination: '/mes-jeunes',
-      permanent: false,
-    },
-  }
-}
-
-export default withTransaction(Milo.name, 'page')(Milo)
